add changes to mpi_perf_tests

+++ tasks/mpi/rezantseva_a_vector_dot_product/perf_tests/main.cpp
@@ -26,7 +26,7 @@ TEST(rezantseva_a_vector_dot_product_mpi, test_pipeline_run) {
   std::vector<int> v1 = createRandomVector(count_size_vector);
   std::vector<int> v2 = createRandomVector(count_size_vector);
 
-  std::vector<int32_t> res(1, 0);
+  std::vector<int64_t> res(1, 0);
   global_vec = {v1, v2};
   // Create TaskData
   std::shared_ptr<ppc::core::TaskData> taskDataPar = std::make_shared<ppc::core::TaskData>();
@@ -45,7 +45,9 @@ TEST(rezantseva_a_vector_dot_product_mpi, test_pipeline_run) {
   ASSERT_EQ(testMpiTaskParallel->validation(), true);
   testMpiTaskParallel->pre_processing();
   testMpiTaskParallel->run();
+
   testMpiTaskParallel->post_processing();
+
   // Create Perf attributes
   auto perfAttr = std::make_shared<ppc::core::PerfAttr>();
   perfAttr->num_running = 10;
@@ -53,7 +55,7 @@ TEST(rezantseva_a_vector_dot_product_mpi, test_pipeline_run) {
   perfAttr->current_timer = [&] { return current_timer.elapsed(); };
   // Create and init perf results
   auto perfResults = std::make_shared<ppc::core::PerfResults>();
-  int answer = rezantseva_a_vector_dot_product_mpi::vectorDotProduct(v1, v2);
+  int64_t answer = rezantseva_a_vector_dot_product_mpi::vectorDotProduct(v1, v2);
   //  Create Perf analyzer
   auto perfAnalyzer = std::make_shared<ppc::core::Perf>(testMpiTaskParallel);
   perfAnalyzer->pipeline_run(perfAttr, perfResults);
@@ -67,7 +69,7 @@ TEST(rezantseva_a_vector_dot_product_mpi, test_pipeline_run) {
 TEST(rezantseva_a_vector_dot_product_mpi, test_task_run) {
   boost::mpi::communicator world;
   std::vector<std::vector<int>> global_vec;
-  std::vector<int32_t> res(1, 0);
+  std::vector<int64_t> res(1, 0);
   std::vector<int> v1 = createRandomVector(count_size_vector);
   std::vector<int> v2 = createRandomVector(count_size_vector);
   // Create TaskData
